requirements.txt
# MangaAutoColor Pro - Dependências (Pinned)
# Generated from working environment, 2026-02-15

# Deep Learning (Core)
# NOTE: PyTorch with CUDA is installed separately.
# For CUDA 12.1: pip install torch==2.5.1+cu121 torchvision==0.20.1+cu121 torchaudio --extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.0.0

# Diffusion / Transformers
diffusers~=0.36.0
transformers~=5.1.0
accelerate~=1.12.0
safetensors~=0.7.0
peft~=0.18.0

# Computer Vision
opencv-python~=4.11.0
opencv-contrib-python~=4.11.0
pillow~=12.1.0
scikit-image~=0.25.0
scipy~=1.15.0
numpy~=1.26.0

# Detection
ultralytics~=8.4.0

# Segmentation (ADR 004)
sam2~=1.1.0

# Vector Database
faiss-cpu~=1.13.0

# API
fastapi~=0.129.0
uvicorn~=0.40.0
pydantic~=2.12.0
python-multipart>=0.0.5

# Templates
jinja2~=3.1.0

# Utilities
tqdm~=4.67.0
colorama>=0.4.0
requests>=2.31.0

# Logging
colorlog>=6.7.0

# Testing (dev)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0
